Updates to Calendar + Frames
This update added functionality of the "submit" button on calendar frame. This allows a user to populate the calendar with any month year combination. Created Try/except blocks for the logic.

Also, I was able to get global dictionary of months to work.

Included changes in the behavior of creating frames. In order to repopulate the calendar with chosen month/year I needed to create a some functionality to redraw the frame inside the calendar.