API communication work
I have continued to work on understanding the use of this api. In the api_choice file I have updated the query result that are display in the box to show a brand name and item name.

In the api_choice_info file I have documented information that helps me sort through the api.