Changes to how modules are working
Changed all of the modules to toplevel classes, so that they run properly.

Made changes to calendar module to open the api_choice module. Old code still there -- Needs cleaning up.

Changed the api_choice module to allow for the passing of the date, month, year, so that I can use that information to create a date title for the window and to allow use in saving the data.